Incorprating comments from Code Walkthrough!
