+++ comid/tdx-profile/example_pce_refval_test.go
@@ -43,38 +43,38 @@ func Example_decode_PCE_JSON() {
 	// Model: 0123456789ABCDEF
 	// InstanceID: 11
 	// pceID: 0000
-	// SVN Operator[0]: 2
-	// SVN Value[0]: 10
-	// SVN Operator[1]: 2
-	// SVN Value[1]: 10
-	// SVN Operator[2]: 2
-	// SVN Value[2]: 2
-	// SVN Operator[3]: 2
-	// SVN Value[3]: 10
-	// SVN Operator[4]: 2
-	// SVN Value[4]: 10
-	// SVN Operator[5]: 2
-	// SVN Value[5]: 10
-	// SVN Operator[6]: 2
-	// SVN Value[6]: 10
-	// SVN Operator[7]: 2
-	// SVN Value[7]: 10
-	// SVN Operator[8]: 2
-	// SVN Value[8]: 10
-	// SVN Operator[9]: 2
-	// SVN Value[9]: 10
-	// SVN Operator[10]: 2
-	// SVN Value[10]: 10
-	// SVN Operator[11]: 2
-	// SVN Value[11]: 10
-	// SVN Operator[12]: 2
-	// SVN Value[12]: 10
-	// SVN Operator[13]: 2
-	// SVN Value[13]: 10
-	// SVN Operator[14]: 2
-	// SVN Value[14]: 10
-	// SVN Operator[15]: 2
-	// SVN Value[15]: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 2
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
 	// CryptoKey Type: pkix-base64-key
 	// CryptoKey Value: -----BEGIN PUBLIC KEY-----
 	// MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEFn0taoAwR3PmrKkYLtAsD9o05KSM6mbgfNCgpuL0g6VpTHkZl73wk5BDxoV7n+Oeee0iIqkW3HMZT3ETiniJdg==
@@ -200,21 +200,8 @@ func extractCompSvn(s *TeeTcbCompSvn) error {
 	}
 
 	for i, teesvn := range *s {
-		if teesvn.IsUint() {
-			svn, err := teesvn.GetUint()
-			if err != nil {
-				return fmt.Errorf("unable to get Uint SVN at index: %d %w", i, err)
-			}
-			fmt.Printf("\nSVN[%d]: %d", i, svn)
-		} else if teesvn.IsNumeric() {
-			svn, err := teesvn.GetNumericExpression()
-			if err != nil {
-				return fmt.Errorf("unable to get SVN Expression at index: %d %w", i, err)
-			}
-			fmt.Printf("\nSVN Operator[%d]: %d", i, svn.NumericOperator)
-			fmt.Printf("\nSVN Value[%d]: %d", i, svn.NumericType.val)
-		} else {
-			return fmt.Errorf("invalid teesvn at index: %d", i)
+		if err := extractTeeSvn(&teesvn); err != nil {
+			return fmt.Errorf("unable to extract SVN at index: %d %w", i, err)
 		}
 	}
 
@@ -259,38 +246,38 @@ func Example_decode_PCE_CBOR() {
 	// Model: TDX PCE TCB
 	// InstanceID: 00112233445566778899aabbccddeeff
 	// pceID: 0000
-	// SVN Operator[0]: 2
-	// SVN Value[0]: 10
-	// SVN Operator[1]: 2
-	// SVN Value[1]: 10
-	// SVN Operator[2]: 2
-	// SVN Value[2]: 2
-	// SVN Operator[3]: 2
-	// SVN Value[3]: 2
-	// SVN Operator[4]: 2
-	// SVN Value[4]: 2
-	// SVN Operator[5]: 2
-	// SVN Value[5]: 1
-	// SVN Operator[6]: 2
-	// SVN Value[6]: 4
-	// SVN Operator[7]: 2
-	// SVN Value[7]: 0
-	// SVN Operator[8]: 2
-	// SVN Value[8]: 0
-	// SVN Operator[9]: 2
-	// SVN Value[9]: 0
-	// SVN Operator[10]: 2
-	// SVN Value[10]: 0
-	// SVN Operator[11]: 2
-	// SVN Value[11]: 0
-	// SVN Operator[12]: 2
-	// SVN Value[12]: 0
-	// SVN Operator[13]: 2
-	// SVN Value[13]: 0
-	// SVN Operator[14]: 2
-	// SVN Value[14]: 0
-	// SVN Operator[15]: 2
-	// SVN Value[15]: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 10
+	// SVN Operator: greater_or_equal
+	// SVN Value: 2
+	// SVN Operator: greater_or_equal
+	// SVN Value: 2
+	// SVN Operator: greater_or_equal
+	// SVN Value: 2
+	// SVN Operator: greater_or_equal
+	// SVN Value: 1
+	// SVN Operator: greater_or_equal
+	// SVN Value: 4
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
+	// SVN Operator: greater_or_equal
+	// SVN Value: 0
 	// CryptoKey Type: pkix-base64-key
 	// CryptoKey Value: -----BEGIN PUBLIC KEY-----
 	// MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEFn0taoAwR3PmrKkYLtAsD9o05KSM6mbgfNCgpuL0g6VpTHkZl73wk5BDxoV7n+Oeee0iIqkW3HMZT3ETiniJdg==
@@ -350,7 +337,7 @@ func Example_encode_tdx_pce_refval_with_profile() {
 
 	// Output:
 	// a301a1005043bbe37f2e614b33aed353cff1428b200281a30065494e54454c01d8207168747470733a2f2f696e74656c2e636f6d028301000204a1008182a100a300d86f4c6086480186f84d01020304050171496e74656c20436f72706f726174696f6e026b544458205043452054434281a101a3384c182d384f685043454944303031387c90d9ea6a820201d9ea6a820202d9ea6a820203d9ea6a820204d9ea6a820205d9ea6a820206d9ea6a820207d9ea6a820208d9ea6a820209d9ea6a82020ad9ea6a82020bd9ea6a82020cd9ea6a82020dd9ea6a82020ed9ea6a82020fd9ea6a820210
-	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDX PCE TCB"}},"measurements":[{"value":{"instanceid":{"type":"uint","value":45},"pceid":"PCEID001","tcbcompsvn":[{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":1}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":2}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":3}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":4}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":5}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":6}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":7}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":8}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":9}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":10}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":11}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":12}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":13}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":14}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":15}}},{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":16}}}]}}]}]}}
+	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDX PCE TCB"}},"measurements":[{"value":{"instanceid":{"type":"uint","value":45},"pceid":"PCEID001","tcbcompsvn":[{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":1}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":2}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":3}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":4}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":5}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":6}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":7}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":8}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":9}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":10}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":11}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":12}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":13}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":14}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":15}}},{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":16}}}]}}]}]}}
 }
 
 func setTDXPCEMvalExtension(val *comid.Mval) error {
@@ -372,7 +359,7 @@ func setTDXPCEMvalExtension(val *comid.Mval) error {
 		return fmt.Errorf("unable to set teepceID %w", err)
 	}
 
-	c, err := NewTeeTcbCompSvnNumeric(TestCompSvn)
+	c, err := NewTeeTcbCompSvnExpression(TestCompSvn)
 	if err != nil {
 		return fmt.Errorf("failed to get TeeTcbCompSvn %w", err)
 	}
