+++ comid/tdx-profile/example_qe_refval_test.go
@@ -128,19 +128,8 @@ func decodeQEMValExtensions(m *comid.Measurement) error {
 		fmt.Printf("val was not pointer to TeeTcbEvalNum")
 	}
 	tcbValNum := *t
-	if tcbValNum.IsNumeric() {
-		ne, err1 := tcbValNum.GetNumericExpression()
-		if err1 != nil {
-			return fmt.Errorf("failed to get tcbEvalNum numeric expression: %w", err1)
-		}
-		fmt.Printf("\ntcbEvalNum Operator: %s", NumericOperatorToString[ne.NumericOperator])
-		fmt.Printf("\ntcbEvalNum Value: %d", ne.NumericType.val)
-	} else if tcbValNum.IsUint() {
-		nv, err1 := tcbValNum.GetUint()
-		if err1 != nil {
-			return fmt.Errorf("failed to get tcbEvalNum uint: %w", err1)
-		}
-		fmt.Printf("\ntcbEvalNum: %d", nv)
+	if err = extractTeeTcbEvalNum(&tcbValNum); err != nil {
+		return fmt.Errorf("failed to extract tcbevalnum: %w", err)
 	}
 
 	val, err = m.Val.Extensions.Get("isvprodid")
@@ -151,23 +140,9 @@ func decodeQEMValExtensions(m *comid.Measurement) error {
 	if !ok {
 		fmt.Printf("val was not pointer to IsvProdID")
 	}
-	if tS == nil {
-		return fmt.Errorf("ISVPRODID IS NIL")
-	}
-	if tS.IsBytes() {
-		val, err = tS.GetBytes()
-		if err != nil {
-			return fmt.Errorf("failed to decode isvprodid: %w", err)
-		}
-		fmt.Printf("\nIsvProdID: %x", val)
-	} else if tS.IsUint() {
-		val, err = tS.GetUint()
-		if err != nil {
-			return fmt.Errorf("failed to decode isvprodid: %w", err)
-		}
-		fmt.Printf("\nIsvProdID: %d", val)
-	} else {
-		return fmt.Errorf("isvprodid is neither integer or byte string")
+
+	if err = extractTeeISVProdID(tS); err != nil {
+		return fmt.Errorf("failed to decode teeISVProdID from measurement extensions: %w", err)
 	}
 
 	val, err = m.Val.Extensions.Get("mrsigner")
@@ -230,11 +205,11 @@ func Example_encode_tdx_QE_refval_without_profile() {
 
 	// Output:
 	// a301a1005043bbe37f2e614b33aed353cff1428b200281a30065494e54454c01d8207168747470733a2f2f696e74656c2e636f6d028301000204a1008182a100a300d86f4c6086480186f84d01020304050171496e74656c20436f72706f726174696f6e02703031323334353637383941424344454681a101a53848d9ea6a82020a385046c000fbff00003853d9ea7482068282015820e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d7582075830e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d75e45b72f5c0c0b572db4d8d3ab7e97f363854013855d9ea6a82020b
-	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"0123456789ABCDEF"}},"measurements":[{"value":{"isvsvn":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":10}}},"miscselect":"wAD7/wAA","mrsigner":{"type":"digest-expression","value":{"set-operator":6,"set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"uint","value":1},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":11}}}}}]}]}}
+	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"0123456789ABCDEF"}},"measurements":[{"value":{"isvsvn":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":10}}},"miscselect":"wAD7/wAA","mrsigner":{"type":"digest-expression","value":{"set-operator":"member","set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"uint","value":1},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":11}}}}}]}]}}
 }
 
 func setTDXQEMvalExtensions(val *comid.Mval) error {
-	svn, err := NewSvnNumeric(TestISVSVN)
+	svn, err := NewSvnExpression(TestISVSVN)
 	if err != nil {
 		return fmt.Errorf("unable to get isvsvn numeric %w", err)
 	}
