+++ comid/tdx-profile/example_seam_refval_test.go
@@ -104,10 +104,12 @@ func Example_encode_tdx_seam_refval_without_profile() {
 	json, err := m.ToJSON()
 	if err == nil {
 		fmt.Printf("%s\n", string(json))
+	} else {
+		fmt.Printf("unable to format json %s, %s", err.Error(), json)
 	}
 	// Output:
 	// a301a1005043bbe37f2e614b33aed353cff1428b200281a30065494e54454c01d8207168747470733a2f2f696e74656c2e636f6d028301000204a1008182a100a300d86f4c6086480186f84d01020304050171496e74656c20436f72706f726174696f6e02675444585345414d81a101a73847c11a6796cc803848d9ea6a82020a38514201013852d9ea7482068182015820e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d753853d9ea7482068282015820e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d7582075830e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d75e45b72f5c0c0b572db4d8d3ab7e97f3638544201013855d9ea6a82020b
-	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDXSEAM"}},"measurements":[{"value":{"tcbdate":"2025-01-27T00:00:00Z","isvsvn":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":10}}},"attributes":"AQE=","mrtee":{"type":"digest-expression","value":{"set-operator":6,"set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU="]}},"mrsigner":{"type":"digest-expression","value":{"set-operator":6,"set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"bytes","value":"AQE="},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":11}}}}}]}]}}
+	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDXSEAM"}},"measurements":[{"value":{"tcbdate":"2025-01-27T00:00:00Z","isvsvn":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":10}}},"attributes":"AQE=","mrtee":{"type":"digest-expression","value":{"set-operator":"member","set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU="]}},"mrsigner":{"type":"digest-expression","value":{"set-operator":"member","set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"bytes","value":"AQE="},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":11}}}}}]}]}}
 }
 
 func Example_encode_tdx_seam_refval_with_profile() {
@@ -165,7 +167,7 @@ func Example_encode_tdx_seam_refval_with_profile() {
 
 	// Output:
 	// a301a1005043bbe37f2e614b33aed353cff1428b200281a30065494e54454c01d8207168747470733a2f2f696e74656c2e636f6d028301000204a1008182a100a300d86f4c6086480186f84d01020304050171496e74656c20436f72706f726174696f6e02675444585345414d81a101a73847c11a6796cc803848d9ea6a82020a38514201013852d9ea7482068182015820e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d753853d9ea7482068282015820e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d7582075830e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d75e45b72f5c0c0b572db4d8d3ab7e97f3638544201013855d9ea6a82020b
-	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDXSEAM"}},"measurements":[{"value":{"tcbdate":"2025-01-27T00:00:00Z","isvsvn":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":10}}},"attributes":"AQE=","mrtee":{"type":"digest-expression","value":{"set-operator":6,"set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU="]}},"mrsigner":{"type":"digest-expression","value":{"set-operator":6,"set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"bytes","value":"AQE="},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":11}}}}}]}]}}
+	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDXSEAM"}},"measurements":[{"value":{"tcbdate":"2025-01-27T00:00:00Z","isvsvn":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":10}}},"attributes":"AQE=","mrtee":{"type":"digest-expression","value":{"set-operator":"member","set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU="]}},"mrsigner":{"type":"digest-expression","value":{"set-operator":"member","set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"bytes","value":"AQE="},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":11}}}}}]}]}}
 }
 
 func Example_encode_tdx_seam_refval_direct() {
@@ -214,7 +216,7 @@ func Example_encode_tdx_seam_refval_direct() {
 
 	// Output:
 	// a301a1005043bbe37f2e614b33aed353cff1428b200281a30065494e54454c01d8207168747470733a2f2f696e74656c2e636f6d028301000204a1008182a100a300d86f4c6086480186f84d01020304050171496e74656c20436f72706f726174696f6e02675444585345414d81a101a73847c11a6796cc803848d9ea6a82020a38514201013852d9ea7482068182015820e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d753853d9ea7482068282015820e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d7582075830e45b72f5c0c0b572db4d8d3ab7e97f368ff74e62347a824decb67a84e5224d75e45b72f5c0c0b572db4d8d3ab7e97f3638544201013855d9ea6a82020b
-	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDXSEAM"}},"measurements":[{"value":{"tcbdate":"2025-01-27T00:00:00Z","isvsvn":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":10}}},"attributes":"AQE=","mrtee":{"type":"digest-expression","value":{"set-operator":6,"set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU="]}},"mrsigner":{"type":"digest-expression","value":{"set-operator":6,"set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"bytes","value":"AQE="},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":2,"numeric-type":{"type":"uint","value":11}}}}}]}]}}
+	// {"tag-identity":{"id":"43bbe37f-2e61-4b33-aed3-53cff1428b20"},"entities":[{"name":"INTEL","regid":"https://intel.com","roles":["creator","tagCreator","maintainer"]}],"triples":{"reference-values":[{"environment":{"class":{"id":{"type":"oid","value":"2.16.840.1.113741.1.2.3.4.5"},"vendor":"Intel Corporation","model":"TDXSEAM"}},"measurements":[{"value":{"tcbdate":"2025-01-27T00:00:00Z","isvsvn":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":10}}},"attributes":"AQE=","mrtee":{"type":"digest-expression","value":{"set-operator":"member","set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU="]}},"mrsigner":{"type":"digest-expression","value":{"set-operator":"member","set-digest":["sha-256;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU=","sha-384;5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXXkW3L1wMC1cttNjTq36X82"]}},"isvprodid":{"type":"bytes","value":"AQE="},"tcbevalnum":{"type":"numeric-expression","value":{"numeric-operator":"greater_or_equal","numeric-type":{"type":"uint","value":11}}}}}]}]}}
 }
 
 func setTDXSeamMvalExtensions(val *comid.Mval) error {
@@ -232,7 +234,7 @@ func setTDXSeamMvalExtensions(val *comid.Mval) error {
 	if err != nil {
 		return fmt.Errorf("unable to set isvprodid %w", err)
 	}
-	svn, err := NewSvnNumeric(TestISVSVN)
+	svn, err := NewSvnExpression(TestISVSVN)
 	if err != nil {
 		return fmt.Errorf("unable to get isvsvn numeric %w", err)
 	}
@@ -296,19 +298,8 @@ func decodeMValExtensions(m *comid.Measurement) error {
 		fmt.Printf("val was not pointer to TeeTcbEvalNum")
 	}
 	tcbValNum := *f
-	if tcbValNum.IsNumeric() {
-		ne, err1 := tcbValNum.GetNumericExpression()
-		if err1 != nil {
-			return fmt.Errorf("failed to get tcbEvalNum numeric expression: %w", err1)
-		}
-		fmt.Printf("\ntcbEvalNum Operator: %s", NumericOperatorToString[ne.NumericOperator])
-		fmt.Printf("\ntcbEvalNum Value: %d", ne.NumericType.val)
-	} else if tcbValNum.IsUint() {
-		nv, err1 := tcbValNum.GetUint()
-		if err1 != nil {
-			return fmt.Errorf("failed to get tcbEvalNum uint: %w", err1)
-		}
-		fmt.Printf("\ntcbEvalNum: %d", nv)
+	if err = extractTeeTcbEvalNum(&tcbValNum); err != nil {
+		return fmt.Errorf("failed to extract tcbevalnum: %w", err)
 	}
 
 	val, err = m.Val.Extensions.Get("isvprodid")
@@ -319,20 +310,8 @@ func decodeMValExtensions(m *comid.Measurement) error {
 	if !ok {
 		fmt.Printf("val was not pointer to IsvProdID")
 	}
-	if tS.IsBytes() {
-		val, err = tS.GetBytes()
-		if err != nil {
-			return fmt.Errorf("failed to decode isvprodid: %w", err)
-		}
-		fmt.Printf("\nIsvProdID: %x", val)
-	} else if tS.IsUint() {
-		val, err = tS.GetUint()
-		if err != nil {
-			return fmt.Errorf("failed to decode isvprodid: %w", err)
-		}
-		fmt.Printf("\nIsvProdID: %d", val)
-	} else {
-		return fmt.Errorf("isvprodid is neither integer or byte string")
+	if err = extractTeeISVProdID(tS); err != nil {
+		return fmt.Errorf("failed to decode teeISVProdID from measurement extensions: %w", err)
 	}
 
 	val, err = m.Val.Extensions.Get("isvsvn")
@@ -348,23 +327,10 @@ func decodeMValExtensions(m *comid.Measurement) error {
 		return fmt.Errorf("invalid tee svn: %w", err)
 	}
 
-	if teesvn.IsUint() {
-		svn, err1 := teesvn.GetUint()
-		if err1 != nil {
-			return fmt.Errorf("unable to get Uint SVN at index: %w", err)
-		}
-		fmt.Printf("\nISVSVN: %d", svn)
-	} else if teesvn.IsNumeric() {
-		svn, err1 := teesvn.GetNumericExpression()
-		if err1 != nil {
-			return fmt.Errorf("unable to get SVN Expression: %w", err)
-		}
-		fmt.Printf("\nSVN Operator: %s", NumericOperatorToString[svn.NumericOperator])
-		fmt.Printf("\nSVN Value: %d", svn.NumericType.val)
-	} else {
-		return fmt.Errorf("teesvn, is neither uint or numeric")
+	err = extractTeeSvn(teesvn)
+	if err != nil {
+		return fmt.Errorf("unable to extract tee svn: %w", err)
 	}
-
 	val, err = m.Val.Extensions.Get("attributes")
 	if err != nil {
 		return fmt.Errorf("failed to decode attributes from measurement extensions")
@@ -403,15 +369,6 @@ func decodeMValExtensions(m *comid.Measurement) error {
 	if err := extractTeeDigest("mrsigner", tD); err != nil {
 		return fmt.Errorf("failed to extarct mrsigner digest: %w", err)
 	}
-	return nil
-}
-
-func decodeAuthorisedBy(m *comid.Measurement) error {
-	if err := m.AuthorizedBy.Valid(); err != nil {
-		return fmt.Errorf("invalid cryptokey: %w", err)
-	}
-	fmt.Printf("\nCryptoKey Type: %s", m.AuthorizedBy.Type())
-	fmt.Printf("\nCryptoKey Value: %s", m.AuthorizedBy.String())
 	return nil
 }
 
@@ -515,82 +472,6 @@ func extractSeamMeasurements(meas *comid.Measurements) error {
 			if err != nil {
 				return fmt.Errorf("extracting measurement at index %d: %w", i, err)
 			}
-		}
-	}
-	return nil
-}
-
-func extractClassElements(c *comid.Class) error {
-	if c == nil {
-		return fmt.Errorf("no class")
-	}
-
-	classID := c.ClassID
-
-	if classID == nil {
-		return fmt.Errorf("no class-id")
-	}
-
-	if classID.Type() != comid.OIDType {
-		return fmt.Errorf("class id is not an oid")
-	}
-
-	fmt.Printf("OID: %s", classID.Value.String())
-
-	if c.Vendor == nil {
-		return fmt.Errorf("no Vendor")
-	}
-	fmt.Printf("\nVendor: %s", *c.Vendor)
-
-	if c.Model == nil {
-		return fmt.Errorf("no Model")
-	}
-	fmt.Printf("\nModel: %s", *c.Model)
-
-	return nil
-}
-
-func extractDigest(typ string, d *Digests) error {
-	if d == nil {
-		return fmt.Errorf("no digest")
-	}
-
-	for _, digest := range *d {
-		fmt.Printf("\n%s Digest Alg: %d", typ, digest.HashAlgID)
-		fmt.Printf("\n%s Digest Value: %x", typ, digest.HashValue)
-	}
-
-	return nil
-}
-
-func extractTeeDigest(typ string, d *TeeDigest) error {
-	if d == nil {
-		return fmt.Errorf("no TEE digest")
-	}
-
-	if typ != "mrsigner" && typ != "mrtee" {
-		return fmt.Errorf("invalid type for TEE digest: %s", typ)
-	}
-
-	if d.IsDigests() {
-		dg, err := d.GetDigest()
-		if err != nil {
-			return fmt.Errorf("unable to extract TEE Digest: %w", err)
-		}
-		err = extractDigest(typ, &dg)
-		if err != nil {
-			return fmt.Errorf("unable to extract %s Digest: %w", typ, err)
-		}
-	} else {
-		de, err := d.GetDigestExpr()
-		if err != nil {
-			return fmt.Errorf("unable to extract TEE Digest Expression: %w", err)
-		}
-		fmt.Printf("\n%s Digest Operator: %s", typ, NumericOperatorToString[de.SetOperator])
-		dg := comid.Digests(de.SetDigest)
-		err = extractDigest(typ, &dg)
-		if err != nil {
-			return fmt.Errorf("unable to extract %s Digest: %w", typ, err)
 		}
 	}
 	return nil
